Update euler_phi (欧拉函数) O (sqrt (n)).cpp

+++ c++/模板/数学/数论/euler_phi (欧拉函数) O (sqrt (n)).cpp
@@ -1,6 +1,6 @@
 // 1 : 
-auto get_euler_phi = [&](int n) -> int {
-    int ans = n;
+auto get_euler_phi = [&](ll n) -> ll {
+    ll ans = n;
     for (int i = 2; i * i <= n; i++) if (n % i == 0) {
         ans = ans / i * (i - 1);
         while (n % i == 0) n /= i;
@@ -10,8 +10,8 @@ auto get_euler_phi = [&](int n) -> int {
 };
 
 // 2 : prime
-auto get_euler_phi = [&](int n) -> int {
-    int ans = n;
+auto get_euler_phi = [&](ll n) -> ll {
+    ll ans = n;
     for (auto v : prime) {
         if (v * v > n) break;
         if (n % v == 0) ans -= ans / v;
